Final Updates
Move components up 1 file level.
Add a carousel for tokens since the list will get long.
Clean up stylings.
Edit some error handling.
