+++ src/components/Footer/styled.ts
