+++ src/components/GetPrivacy/questions.ts
