+++ src/components/GetPrivacy/styled.ts
