+++ src/components/Loader/styled.ts
