+++ src/components/TokenForm/Indicators/styled.ts
