+++ src/components/TokenForm/WithdrawIBC/WithdrawIBC.tsx
@@ -1,38 +1,47 @@
-import { DropDownMenu } from "../../../DropDownMenu/DropDownMenu";
-import { Token, Chain } from "../../../../types";
-import { ChainList } from "../../../../config";
+import { ChainList } from "../../../config";
+import { Token, Chain } from "../../../types";
+import { SecretNetworkClient } from "secretjs";
+import { DropDownMenu } from "../../DropDownMenu/DropDownMenu";
 import { CopyableAddress } from "../CopyableAddress/CopyableAddress";
-import { Button } from "../../../Button/Button";
-import { useEffect, useRef, useState } from "react";
 import { SigningStargateClient } from "@cosmjs/stargate";
-import { depositTx } from "./depositTx";
-import { getIBCBalance } from "../../Helpers/data";
+import { Button } from "../../Button/Button";
+import { useEffect, useRef, useState } from "react";
 import { fixedBalance } from "../../Helpers/format";
+import { getTokenBalance } from "../../Helpers/data";
 import { setupCosmjs } from "../../Helpers/cosm";
+import { withdrawTx } from "./withdrawTx";
 import { Loader } from "../../Loader/Loader";
 
-interface DepositProps {
-  secretAddress: string;
+interface WithdrawProps {
   currentToken: Token;
+  secretjs: SecretNetworkClient | null;
+  secretAddress: string;
 }
 
-export const DepositIBC = ({ secretAddress, currentToken }: DepositProps) => {
+export const WithdrawIBC = ({
+  currentToken,
+  secretjs,
+  secretAddress,
+}: WithdrawProps) => {
   const [addressIBC, setAdressIBC] = useState<string>("");
   const [balanceIBC, setBalanceIBC] = useState<string>("0");
+  const [tokenBalance, setTokenBalance] = useState<string>("0");
 
-  const [loadingDeposit, setLoadingDeposit] = useState<boolean>(false);
-  const [loadingBalanceIBC, setLoadingBalanceIBC] = useState<boolean>(false);
-
+  const [loadingWithdrawal, setLoadingWithdrawal] = useState<boolean>(false);
+  const [loadingTokenBalance, setLoadingTokenBalance] =
+    useState<boolean>(false);
   const [cosmjs, setCosmjs] = useState<SigningStargateClient | null>(null);
   const [selectedChainIndex, setSelectedChainIndex] = useState<number>(0);
+
+  const inputRef = useRef<any>();
 
   const [targetChain, setTargetChain] = useState<Chain>(
     ChainList[currentToken.deposits[0].source_chain_name]
   );
-  const inputRef = useRef<any>();
 
   useEffect(() => {
     setBalanceIBC("0");
+    currentToken;
     if (currentToken.name !== "SCRT") {
       setSelectedChainIndex(0);
       setupCosmjs(
@@ -40,6 +49,12 @@ export const DepositIBC = ({ secretAddress, currentToken }: DepositProps) => {
         setAdressIBC,
         ChainList[currentToken.deposits[0].source_chain_name],
         currentToken
+      );
+      getTokenBalance(
+        currentToken,
+        secretAddress,
+        setTokenBalance,
+        setLoadingTokenBalance
       );
     } else {
       setupCosmjs(
@@ -55,55 +70,49 @@ export const DepositIBC = ({ secretAddress, currentToken }: DepositProps) => {
   }, [currentToken, targetChain]);
 
   useEffect(() => {
-    if (addressIBC) {
-      getIBCBalance(
-        addressIBC,
-        currentToken,
-        selectedChainIndex,
-        setBalanceIBC,
-        setLoadingBalanceIBC
-      );
-    }
+    getTokenBalance(
+      currentToken,
+      secretAddress,
+      setTokenBalance,
+      setLoadingTokenBalance
+    );
   }, [addressIBC]);
 
   return (
     <div style={{ display: "flex", flexDirection: "column" }}>
       <div className="deposit-block">
-        <p>Deposit {currentToken.name} from</p>
+        <p>Withdraw {currentToken.name} to</p>
         {currentToken.name === "SCRT" ? (
           <DropDownMenu
             currentToken={currentToken}
             selectedChainIndex={selectedChainIndex}
             setSelectedChainIndex={setSelectedChainIndex}
             setTargetChain={setTargetChain}
-            tab={"deposits"}
+            tab={"withdrawals"}
           />
         ) : (
-          <span className="one">
-            {ChainList[currentToken.deposits[0].source_chain_name].chain_name}
-          </span>
+          <span className="one">{targetChain!.chain_name}</span>
         )}
       </div>
 
       <CopyableAddress
         title="From"
-        address={addressIBC}
-        prefix={targetChain!.explorer_account}
-      />
-      <CopyableAddress
-        title="To"
         address={secretAddress}
         prefix={ChainList["Secret Network"].explorer_account}
       />
-
+      <CopyableAddress
+        title="To"
+        address={addressIBC}
+        prefix={targetChain!.explorer_account}
+      />
       <div className="available">
         <span className="title">Available Balance:</span>
         <span className="cash">
-          {loadingBalanceIBC ? (
+          {loadingTokenBalance ? (
             <Loader />
           ) : (
-            <span className="available-deposit">
-              {`${fixedBalance(balanceIBC, currentToken.decimals)} ${
+            <span className="available-withdrawal">
+              {`${fixedBalance(tokenBalance, currentToken.decimals)} ${
                 currentToken.name
               }`}
             </span>
@@ -112,28 +121,28 @@ export const DepositIBC = ({ secretAddress, currentToken }: DepositProps) => {
       </div>
 
       <div className="amount">
-        <input ref={inputRef} placeholder="Amount to Deposit" />
+        <input ref={inputRef} placeholder="Amount to Withdraw" />
         <img src={currentToken.image} alt="amount" />
       </div>
 
       <Button
         action={() => {
           currentToken.name === "SCRT"
-            ? depositTx(
-                cosmjs,
+            ? withdrawTx(
+                secretjs,
                 inputRef,
                 targetChain,
                 addressIBC,
                 secretAddress,
                 currentToken,
                 selectedChainIndex,
-                loadingDeposit,
-                setLoadingDeposit,
-                setBalanceIBC,
-                setLoadingBalanceIBC
+                loadingWithdrawal,
+                setLoadingWithdrawal,
+                setTokenBalance,
+                setLoadingTokenBalance
               )
-            : depositTx(
-                cosmjs,
+            : withdrawTx(
+                secretjs,
                 inputRef,
                 ChainList[
                   currentToken.deposits[selectedChainIndex].source_chain_name
@@ -142,14 +151,14 @@ export const DepositIBC = ({ secretAddress, currentToken }: DepositProps) => {
                 secretAddress,
                 currentToken,
                 selectedChainIndex,
-                loadingDeposit,
-                setLoadingDeposit,
-                setBalanceIBC,
-                setLoadingBalanceIBC
+                loadingWithdrawal,
+                setLoadingWithdrawal,
+                setTokenBalance,
+                setLoadingTokenBalance
               );
         }}
-        text={"deposit"}
-        isLoading={loadingDeposit}
+        text={"withdraw"}
+        isLoading={loadingWithdrawal}
       />
     </div>
   );
