+++ vite.config.ts
@@ -1,6 +1,9 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
 
 export default defineConfig({
-  plugins: [react()]
-})
+  plugins: [react()],
+  esbuild: {
+    logOverride: { "this-is-undefined-in-esm": "silent" },
+  },
+});
